pyproject.toml
[build-system]
requires = [
    "setuptools>=61.0",
    "wheel",
    "build"
    ]
build-backend = "setuptools.build_meta"

[project]
name = "smia"   # cambiar cuando funcione
version = '0.2.1'  # cambiar cuando funcione a 0.2.2
authors = [{ name="Ekaitz Hurtado", email="ekaitz.hurtado@ehu.eus" },]
description = "Self-configurable Manufacturing Industrial Agent (SMIA)."
readme = 'README_pypi.md'
license= {text="GNU Affero General Public License v3"}
keywords=['smia', 'dt', 'aas', 'i4.0']
requires-python = ">= 3.10"
classifiers = [
    'Development Status :: 2 - Pre-Alpha',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: GNU Affero General Public License v3',
    'Natural Language :: English',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12',
]
dependencies = [
    "spade == 3.3.3",
    "basyx-python-sdk == 1.1.0",
    "owlready2 == 0.47",
    "jsonpath-ng == 1.7.0",
    "aiohttp >= 3.10.5",
    "psutil == 6.1.0",
    "jsonschema == 4.22.0",
    "tqdm ~= 4.67.1"
]

[project.urls]
Homepage = "https://github.com/ekhurtado/I4_0_SMIA"
Documentation = "https://i4-0-smia.readthedocs.io/en/latest/"
Repository = "https://github.com/ekhurtado/I4_0_SMIA"
Issues = "https://github.com/ekhurtado/I4_0_SMIA/issues"
Changelog = "https://i4-0-smia.readthedocs.io/en/latest/release_notes.html"

#[tool.setuptools.dynamic]
#dependencies = {file = "docker/requirements.txt"}  # This is a beta feature